fix: feature scope
